カットインのバリエーション追加のためディレクトリ構成を変更
